Make tries a u32
Before this variable was an u8, which could theoretically result in desyncs with a large buffer(greater than 255*120/200=153) filled with blank inputs. If this could actually happen, i don't know. But this part of the code on its own looks like it could break.

+++ Source/Core/Core/NetPlayClient.cpp
@@ -989,7 +989,7 @@ bool NetPlayClient::WiimoteUpdate(int _number, u8* data, const u8 size)
   // until we reach a good input
   if (nw.size() != size)
   {
-    u8 tries = 0;
+    u32 tries = 0;
     // Clear the buffer and wait for new input, since we probably just changed reporting mode.
     while (nw.size() != size)
     {
